move generated folder
