+++ openapi-ts.config.ts
@@ -2,7 +2,7 @@ import { defaultPlugins } from '@hey-api/openapi-ts';
 
 export default {
   input: 'http://localhost:3147/doc',
-  output: './packages/client/src/hooks/generated',
+  output: './packages/client/src/generated',
   plugins: [
     ...defaultPlugins,
     '@hey-api/client-fetch',
