+++ packages/client/src/components/navigation/file-viewer-dialog.tsx
@@ -8,7 +8,7 @@ import { MarkdownRenderer } from '@/components/markdown-renderer'
 import { useCopyClipboard } from '@/hooks/utility-hooks/use-copy-clipboard'
 import { Switch } from '@ui'
 import { useSelectSetting } from '@/hooks/api/use-kv-api'
-import { ProjectFile } from '@/hooks/generated'
+import { ProjectFile } from '@/generated'
 import * as themes from "react-syntax-highlighter/dist/esm/styles/hljs"
