+++ packages/client/src/components/projects/project-dialog.tsx
@@ -12,7 +12,7 @@ import { Input } from "@ui"
 import { Label } from "@ui"
 import { useCreateProject, useUpdateProject, useGetProject, useSyncProject } from "@/hooks/api/use-projects-api"
 import { useEffect, useState } from "react"
-import { CreateProjectRequestBody } from "@/hooks/generated"
+import { CreateProjectRequestBody } from "@/generated"
 import { useUpdateActiveProjectTab } from "@/hooks/api/use-kv-api"
 
 type ProjectDialogProps = {
