+++ packages/client/src/components/projects/prompt-overview-panel.tsx
@@ -22,7 +22,7 @@ import { z } from 'zod'
 import { SuggestedFilesDialog } from '../suggest-files-dialog'
 import { VerticalResizablePanel } from '@ui'
 import { useSuggestFiles } from '@/hooks/api/use-gen-ai-api'
-import { ProjectFile } from '@/hooks/generated'
+import { ProjectFile } from '@/generated'
 import { useCreateChat } from '@/hooks/api/use-chat-api'
 import { useLocalStorage } from '@/hooks/utility-hooks/use-local-storage'
 import { Binoculars, Bot, CheckCircle, Copy, MessageCircleCode, RefreshCw, Search, Trash2, Activity } from 'lucide-react'
