+++ packages/client/src/components/projects/prompts-list.tsx
@@ -20,7 +20,7 @@ import { DotsHorizontalIcon } from '@radix-ui/react-icons'
 import { Badge } from '@ui'
 import { InfoTooltip } from '../info-tooltip'
 import { ShortcutDisplay } from '../app-shortcut-display'
-import { ProjectFile } from '@/hooks/generated'
+import { ProjectFile } from '@/generated'
 import { promptSchema } from 'shared/src/utils/projects-utils'
 import { useGetProjectTabById, useUpdateProjectTabState } from '@/hooks/api/use-kv-api'
 import { useCopyClipboard } from '@/hooks/utility-hooks/use-copy-clipboard'
