+++ packages/client/src/components/projects/selected-files-drawer.tsx
@@ -13,7 +13,7 @@ import { useState } from "react"
 import { FormatTokenCount } from "../format-token-count"
 import { Badge } from "@ui"
 import { useSelectedFiles, type UseSelectedFileReturn } from '@/hooks/utility-hooks/use-selected-files'
-import { ProjectFile } from "@/hooks/generated"
+import { ProjectFile } from "@/generated"
 import { ProjectFileMap } from "shared/src/schemas/project.schemas"
 import { estimateTokenCount } from "shared/src/utils/file-tree-utils/file-node-tree-utils"
