+++ packages/client/src/components/suggest-files-dialog.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react"
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription, DialogFooter } from "@ui"
 import { Button } from '@ui'
 import { useSelectedFiles } from "@/hooks/utility-hooks/use-selected-files"
-import { ProjectFile } from "@/hooks/generated"
+import { ProjectFile } from "@/generated"
 
 type SuggestedFilesDialogProps = {
     open: boolean
