+++ packages/client/src/hooks/api/use-keys-api.ts
@@ -6,7 +6,7 @@ import {
     postApiKeysMutation,
     patchApiKeysByKeyIdMutation,
     deleteApiKeysByKeyIdMutation,
-} from '../generated/@tanstack/react-query.gen';
+} from '../../generated/@tanstack/react-query.gen';
 import type {
     ProviderKeyListResponse,
     ProviderKeyResponse,
@@ -18,8 +18,8 @@ import type {
     DeleteApiKeysByKeyIdError,
     DeleteApiKeysByKeyIdResponse,
     ProviderKey,
-} from '../generated/types.gen';
-import { Options } from '../generated/sdk.gen';
+} from '../../generated/types.gen';
+import { Options } from '../../generated/sdk.gen';
 
 export type CreateKeyInput = PostApiKeysData['body'];
 export type UpdateKeyInput = PatchApiKeysByKeyIdData['body'];
