+++ packages/client/src/hooks/api/use-tickets-api.ts
@@ -23,7 +23,7 @@ import {
     postApiTicketsByTicketIdLinkFilesMutation,
     postApiTicketsByTicketIdSuggestFilesMutation,
     postApiTicketsByTicketIdSuggestTasksMutation
-} from '../generated/@tanstack/react-query.gen';
+} from '../../generated/@tanstack/react-query.gen';
 import type {
     CreateTicketBody,
     UpdateTicketBody,
@@ -56,8 +56,8 @@ import type {
     PostApiTicketsByTicketIdSuggestTasksError,
     PostApiTicketsData,
     PostApiTicketsError
-} from '../generated/types.gen';
-import { Options } from '../generated/sdk.gen';
+} from '../../generated/types.gen';
+import { Options } from '../../generated/sdk.gen';
 
 const TICKET_KEYS = {
     all: ['tickets'] as const,
