+++ packages/client/src/hooks/use-ai-chat.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect } from 'react';
 import { useChat, Message } from '@ai-sdk/react';
-import type { AiChatStreamRequest, AiSdkOptions, } from './generated';
+import type { AiChatStreamRequest, AiSdkOptions, } from '../generated';
 import { useGetMessages } from './api/use-chat-api';
 import { APIProviders } from 'shared/src/schemas/provider-key.schemas';
 import { nanoid } from 'nanoid';
