+++ packages/shared/src/utils/file-tree-utils/import-resolve.test.tsx
@@ -1,4 +1,4 @@
-import type { ProjectFile } from '../../../../client/src/hooks/generated'
+import type { ProjectFile } from '../../../../client/src/generated'
 import { gatherAliasesFromTsconfigs, getRecursiveImports, buildTsconfigAliasMap, type TsconfigCache } from './import-resolver'
 import { describe, it, expect, beforeAll } from 'bun:test'
